Utils func reorganized inside profiles package
