+++ config/samples/sw.kogito_v1alpha08_kogitoserverlessworkflow_prodProfile.yaml
@@ -5,7 +5,7 @@ metadata:
   annotations:
     sw.kogito.kie.org/description: Greeting example on k8s!
     sw.kogito.kie.org/version: 0.0.1
-    sw.kogito.kie.org/profile: prod
+    sw.kogito.kie.org/profile: prod #It is not mandatory to add this annotation to use the Prod Profile
 spec:
   start: ChooseOnLanguage
   functions:
